chore: update file structures, move to src
